feature: adding v6 player toggle and refactor old component structure to make easier for v6 player upgrade
