+++ src/components/old-v5-player/loading-tune.js
@@ -4,7 +4,7 @@
  * Copyright zulucoda - mfbproject
  */
 import React from 'react';
-import { LoadingWrapper } from '../utils/styles';
+import { LoadingWrapper } from '../../utils/styles';
 
 export const LoadingTune = () => (
   <LoadingWrapper>
