+++ src/components/old-v5-player/tune-error.js
@@ -4,7 +4,7 @@
  * Copyright zulucoda - mfbproject
  */
 import React from 'react';
-import { WarningText, WarningWrapper } from '../utils/styles';
+import { WarningText, WarningWrapper } from '../../utils/styles';
 
 const TuneError = ({ currentTune }) => (
   <WarningWrapper>
