+++ src/utils/constants.js
@@ -2,3 +2,6 @@ export const HIGH_VOLUME = 1;
 export const MEDIUM_VOLUME = 0.5;
 export const LOW_VOLUME = 0.3;
 export const MUTE_VOLUME = 0;
+
+// Toggle between old and new player
+export const FEATURE_TOGGLE_PLAYER_V6 = true;
